+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/CollectionExtensions.kt
@@ -296,6 +296,18 @@ fun <K, V> Map<K, V>.joinToStringOrEmpty(separator: CharSequence = ", ", prefix:
 }
 
 
+/**过滤映射中值为null的键值对。*/
+inline fun <K, V : Any> Map<K, V?>.filterValueNotNull(): Map<K, V> {
+	return this.filterValueNotNullTo(HashMap())
+}
+
+/**过滤映射中值为null的键值对，然后置入指定的映射。*/
+inline fun <K, V : Any, M : MutableMap<in K, in V>> Map<out K, V?>.filterValueNotNullTo(destination: M): M {
+	for((key, value) in this) if(value != null) destination[key] = value
+	return destination
+}
+
+
 /**按照类型以及附加条件过滤数组。*/
 inline fun <reified R> Array<*>.filterIsInstance(predicate: (R) -> Boolean): List<R> {
 	return this.filterIsInstanceTo<R, MutableList<R>>(ArrayList(), predicate)