+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/DataClassExtensions.kt
@@ -7,10 +7,10 @@ import kotlin.reflect.*
 //NOTE 为了避免污染Any?的代码提示，不作为扩展方法
 //DONE 特殊对待数组类型
 //DONE toString时可以省略值为null的属性
-//TODO 考虑父类
+//DONE toString时可以考虑父类型的toString
 
 /**通过选择并比较指定类型中的属性，判断两个对象是否相等。*/
-inline fun <reified T> equalsBySelect(target: T?, other: Any?, selector: T.() -> Array<*>): Boolean {
+inline fun <reified T> equalsBySelect(target: T?, other: Any?, selector: T .() -> Array<*>): Boolean {
 	if(target === other) return true
 	if(target == null) return false
 	if(other !is T) return false
@@ -33,7 +33,7 @@ inline fun <reified T> toStringBySelect(target: T?, omitNulls: Boolean = false,
 	if(target == null) return "null"
 	val className = T::class.java.simpleName
 	val propertiesSnippet = target.selector().toMap()
-		.where(omitNulls) { it.filter { (_, v) -> v == null } } //TODO
+		.where(omitNulls) { it.filterValueNotNull() }
 		.mapValues { (_, v) ->
 			//treat Arrays specially
 			if(v is Array<*>) v.contentDeepToString() else v.toString()
@@ -42,15 +42,16 @@ inline fun <reified T> toStringBySelect(target: T?, omitNulls: Boolean = false,
 }
 
 /**通过选择指定类型中的属性引用，将指定对象转化为字符串。*/
-inline fun <reified T> toStringBySelectRef(target: T?, omitNulls: Boolean = false,
+inline fun <reified T> toStringBySelectRef(target: T?, omitNulls: Boolean = false, superToString: String? = null,
 	selector: T.() -> Array<KProperty0<*>>): String {
 	if(target == null) return "null"
 	val className = T::class.java.simpleName
 	val propertiesSnippet = target.selector().associate { it.name to it.get() }
-		.where(omitNulls) { it.filter { (_, v) -> v == null } } //TODO
+		.where(omitNulls) { it.filterValueNotNull() }
 		.mapValues { (_, v) ->
 			//treat Arrays specially
 			if(v is Array<*>) v.contentDeepToString() else v.toString()
 		}.joinToString()
-	return "$className($propertiesSnippet)"
+	val superToStringSnippet = superToString?.let { ", $it" }.orEmpty()
+	return "$className($propertiesSnippet$superToStringSnippet)"
 }