+++ CHANGELOG.md
@@ -143,6 +143,7 @@
 * [X] breeze-core 提供标准颜色的枚举。（考虑改为常量）
 * [X] 更新README。
 * [X] breeze-core 优化`toStringBySelect`等方法。
+* [X] breeze-core 提供`filterValueNotNull`、`filterValueNotNullTo`扩展。
 * [ ] breeze-dsl 实现`YamlDsl`。（不推荐使用，但是作为规范）
 
 ***